Video decoding from mmapped buffer, cabac hook.
Rewrite recode.cpp in clean C++.

Use AVIO hook for read_packet to read input data out of an mmapped
buffer.

Add a printf to ff_init_cabac_decoder. Validate that for my test file
GOPR4542.MP4, over 95% of the file bytes are cabac encoded regions.
We'll recode the cabac and store the rest in a side stream.