Use surrogate blocks to replace cabac-encoded bit blocks.
Instead of substituting in CABAC blocks as literal blocks, substitute in
a placeholder "surrogate" block when reading in the input stream in the
decompressor. Then, when the CABAC decoder is initialized, copy in the
real CABAC data.

The next step is to hook in to the get_cabac call on a bit-by-bit basis,
and stop copying in the CABAC data entirely.