Refactor CABAC decoding hooks and improve compressed proto.
Separate the CABAC hooks into their own object, so that coded block
decoding is independent of its harness. In principle, this means that
frames could be decoded in parallel; it also opens the path to having
multiple different coded block handlers.

Improve the compressed file format. Record skipped coded blocks.

Clean up the surrogate marker system so that we don't have as many casts
and we don't have to check the input data for edge cases.