Fixed bugs in CABAC encoding.
Finally, can re-encode all the CABAC blocks in my test file!
The bug that took forever to find:
  return 2 << normalize;
With FixedPoint == uint64_t, `2` was implicitly int, and 2 << normalize
was shifting away to zero. No compiler warning! Yikes.

Also added overflow digits with carry, which turned out to be
surprisingly easy.

Added a bunch of documentation about the representation invariant, which
should turn out handy later.

Refactored the stop bit computation so that CABAC and the generic
arithmetic coder can share the stop bit logic. This version is better
- the old version wasn't wrong per se, but this version's stop bit
more faithfully indicates the precision of the encoded number.
(Still not 100% faithful - it can be longer than necessary. It would be
possible to be 100% faithful but not without losing CABAC
compatibility.)