Store decoded symbols in recoded file, first roundtrip!
Very hacky patch with a lot of #if/#else blocks to replace the
CABAC-encoded block in the recoded file with the raw symbols.
I had to implement bit-packing (which was difficult to debug) because
protobuf can't handle huge files.

This encoding demonstrates the technique working end-to-end. Now all
that remains is to encode the symbol stream efficiently.