Fix build instructions.
The whole ffmpeg package needs to be built the first time because
libavcodec depends on the other libraries in ffmpeg. If only libavcodec
is built, then linking picks up the system libraries, which can mismatch
the libavcodec version.

Once the other libraries are built once, it's only necessary to make
libavcodec/libavcodec.a when the libavcodec source is changed.